boorsocorona_incomplete/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.18
django-decouple==2.1
django-redis==4.12.1
djangorestframework==3.12.2
frozenlist==1.3.3
gunicorn==20.0.4
idna==3.4
multidict==6.0.4
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
yarl==1.8.2
